+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [UNRELEASED]
+### Added
+- Add interface to download NSRR data ([#X](https://github.com/cbrnr/sleepecg/pull/X) by [Florian Hofer](https://github.com/hofaflo))
 ### Fixed
 - Fix MemoryError in case of invalid checksum ([#3](https://github.com/cbrnr/sleepecg/pull/3))